Move fill_protocol_info to obs_data_helpers, where its vocabulary lives
It was put beside the .mmt reader because that is what first needed it, and
then moved out of this package entirely when CUFLynx's unit tier failed on it.
Both were wrong, and the second was a fix for the wrong cause: that tier was
pinned at a commit predating the reader, which is why nothing engine-side could
answer it. The pin is bumped now.

Every key it writes -- protocol_info, data_items, experiment_labels,
experiment_colors -- is obs_data_helpers' vocabulary, and obs_data_helpers is
where those names get migrated when they change (migrate_legacy_obs_item_keys,
issue #466 being the recent one). A copy of this in a downstream repo would
keep writing the old spelling with nothing to catch it.

It is not Myokit-specific either: the EasyML reader produces a protocol_info
too, and so could anything else. Living beside one producer of them was the
original mistake.
